requirements.txt
matplotlib
pandas
numpy
gspread
seaborn
python-dotenv